makedep: Support externals alongside program units
The current version of makedep uses the existence of program units
(module or program) to determine whether a file contains globally
accessible external functions.  It incorrectly ignores files with
externals it happens to contain both externals and a program unit.

We resolve this by tracking whether we are inside or outside of a
program unit while parsing each line.  We ignore keywords which may
appear in comments, as well as tokens which may contain `function` or
`subroutine` as substrings.  This method is not perfect; there are cases
which will incorrectly report entering a block, but this method appears
sufficient for all of our known codebases.

This issue was detected when attempting to parse the AM2 source.  While
still far from a general solution, it is sufficient to handle a few of
the more challenging cases encountered in AM2.